WIP gossip stuff needs to be updated for private field access
