+++ Cargo.toml
@@ -14,16 +14,32 @@ resolver = "2"
 overflow-checks = true
 
 [workspace.dependencies]
-anchor-lang = "0.30.1"
+anchor-lang = "0.31.1"
 axum = "0.6.2"
 clap = { version = "4.3.0", features = ["derive", "env"] }
 http = { version = "0.2.1" }
+jito-priority-fee-distribution = { features = ["no-entrypoint"], git = "https://github.com/jito-foundation/jito-programs", rev = "24ec5d8a170ac86412167056df8aa4af9b12667c" }
+jito-tip-distribution = { features = ["no-entrypoint"], git = "https://github.com/jito-foundation/jito-programs", rev = "24ec5d8a170ac86412167056df8aa4af9b12667c" }
+jito-tip-distribution-sdk = { git = "https://github.com/jito-foundation/jito-programs", rev = "24ec5d8a170ac86412167056df8aa4af9b12667c" }
 serde = "1.0.183"
 serde_derive = "1.0.183"
 serde_json = "1.0.102"
-solana-program = "1.18"
-solana-rpc-client = "1.18"
-solana-rpc-client-api = "1.18"
+solana-bincode = "=2.2.1"
+solana-program = "=2.2.1"
+solana-rpc-client = "=2.2.1"
+solana-rpc-client-api = "=2.2.1"
+solana-account-decoder = "=2.2.1"
+solana-clap-utils = "=2.2.1"
+solana-client = "=2.2.1"
+solana-gossip = "=2.2.1"
+solana-metrics = "=2.2.1"
+solana-net-utils = "=2.2.1"
+solana-packet = "=2.2.1"
+solana-sdk = "=2.2.1"
+solana-streamer = "=2.2.1"
+solana-transaction-status = "=2.2.1"
+spl-stake-pool = { features = ["no-entrypoint"], version = "2.0.2" }
+spl-pod = "0.5.1"
 stakenet-sdk = { path = "sdk", version = "0.1.0" }
 thiserror = "1.0.37"
 tokio = { version = "1.36.0", features = ["full"] }
