+++ keepers/stakenet-keeper/src/operations/gossip_upload.rs
@@ -10,8 +10,9 @@ use bytemuck::{bytes_of, Pod, Zeroable};
 use log::*;
 use solana_client::nonblocking::rpc_client::RpcClient;
 use solana_client::rpc_response::RpcVoteAccountInfo;
+use solana_gossip::crds_data::CrdsData;
 use solana_gossip::crds::Crds;
-use solana_gossip::crds_value::{CrdsData, CrdsValue, CrdsValueLabel};
+use solana_gossip::crds_value::{CrdsValue, CrdsValueLabel};
 use solana_gossip::gossip_service::make_gossip_node;
 use solana_metrics::datapoint_error;
 use solana_sdk::signature::Signable;
