+++ programs/steward/Cargo.toml
@@ -22,17 +22,18 @@ custom-heap = []
 idl-build = ["anchor-lang/idl-build", "validator-history/idl-build", "no-entrypoint"]
 
 [dependencies]
-anchor-lang = "0.30.1"
+anchor-lang = "0.31.1"
 bincode = "1.3.3"
 blake3 = "1.3.1"
 borsh = "0.10.0"
+borsh1 = { package = "borsh", version = "1.5.7" }
 bytemuck = { features = ["derive", "min_const_generics"], version = "1.13.1" }
 cfg-if = "1.0.0"
 semver = "1.0.17"
 serde = "1.0.188"
 serde_derive = "1.0.183"
-spl-pod = "0.1.0"
-spl-stake-pool = { features = ["no-entrypoint"], version = "1.0.0" }
+spl-pod = { workspace = true }
+spl-stake-pool = { features = ["no-entrypoint"], workspace = true }
 static_assertions = "1.1.0"
 thiserror = "1.0.37"
 type-layout = "0.2"
