+++ programs/steward/src/instructions/initialize_steward.rs
@@ -1,8 +1,8 @@
 use anchor_lang::{prelude::*, solana_program::program::invoke};
 
 use crate::{
-    constants::MAX_ALLOC_BYTES, utils::deserialize_stake_pool, Config, StewardStateAccount,
-    UpdateParametersArgs, UpdatePriorityFeeParametersArgs,
+    constants::MAX_ALLOC_BYTES, stake_pool_utils::deserialize_stake_pool, Config,
+    StewardStateAccount, UpdateParametersArgs, UpdatePriorityFeeParametersArgs,
 };
 
 #[derive(Accounts)]
