+++ programs/steward/src/instructions/reset_steward_state.rs
@@ -1,8 +1,9 @@
 use crate::{
     constants::{LAMPORT_BALANCE_DEFAULT, MAX_VALIDATORS, SORTED_INDEX_DEFAULT},
     errors::StewardError,
+    stake_pool_utils::deserialize_stake_pool,
     state::{Config, StewardStateAccount},
-    utils::{deserialize_stake_pool, get_config_admin, get_stake_pool_address},
+    utils::{get_config_admin, get_stake_pool_address},
     BitMask, Delegation, StewardStateEnum, STATE_PADDING_0_SIZE,
 };
 use anchor_lang::prelude::*;
