+++ programs/validator-history/Cargo.toml
@@ -25,16 +25,16 @@ custom-heap = []
 idl-build = ["anchor-lang/idl-build", "no-entrypoint"]
 
 [dependencies]
-anchor-lang = "0.30.1"
+anchor-lang = "0.31.1"
 bincode = "1.3.3"
 bytemuck = { version = "1.13.1", features = ["derive", "min_const_generics"] }
 cfg-if = "1.0.0"
-jito-priority-fee-distribution = { features = ["no-entrypoint"], git = "https://github.com/jito-foundation/jito-programs", branch = "tj/priority-fee-distributor" }
-jito-tip-distribution = { features = ["no-entrypoint"], git = "https://github.com/jito-foundation/jito-programs", branch = "tj/priority-fee-distributor" }
+jito-priority-fee-distribution = { features = ["no-entrypoint"], workspace = true }
+jito-tip-distribution = { features = ["no-entrypoint"], workspace = true }
 semver = "1.0.17"
 serde = "1.0.183"
 serde_derive = "1.0.183"
-solana-program = "1.18"
+solana-program = { workspace = true }
 solana-security-txt = "1.1.0"
 static_assertions = "1.1.0"
 thiserror = "1.0.37"
