+++ sdk/Cargo.toml
@@ -11,21 +11,21 @@ crate-type = ["cdylib", "lib"]
 name = "stakenet_sdk"
 
 [dependencies]
-anchor-lang = "0.30.1"
+anchor-lang = "0.31.1"
 bincode = "1.3.3"
 clap = { version = "4.3.0", features = ["derive"] }
 futures = "0.3.21"
 jito-steward = { features = ["no-entrypoint"], path = "../programs/steward" }
-jito-tip-distribution = { features = ["no-entrypoint"], git = "https://github.com/jito-foundation/jito-programs", rev = "50d450e993cb2278bcf97cd01b19e8a4f1f56e8e" }
+jito-tip-distribution = { features = ["no-entrypoint"], workspace = true }
 log = "0.4.18"
-solana-account-decoder = "1.18"
-solana-client = "1.18"
-solana-metrics = "1.18"
-solana-program = "1.18"
-solana-sdk = "1.18"
-solana-transaction-status = "1.18"
-spl-pod = "0.1.0"
-spl-stake-pool = { features = ["no-entrypoint"], version = "1.0.0" }
+solana-account-decoder = { workspace = true }
+solana-client = { workspace = true }
+solana-metrics = { workspace = true }
+solana-program = { workspace = true }
+solana-sdk = { workspace = true }
+solana-transaction-status = { workspace = true }
+spl-pod = { workspace = true }
+spl-stake-pool = { features = ["no-entrypoint"], workspace = true }
 thiserror = "1.0.37"
 tokio = { version = "1.36.0", features = ["full"] }
 validator-history = { features = ["no-entrypoint"], path = "../programs/validator-history" }
