+++ tests/Cargo.toml
@@ -18,30 +18,31 @@ cpi = ["no-entrypoint"]
 default = []
 
 [dependencies]
-anchor-lang = "0.30.1"
+anchor-lang = "0.31.1"
 bincode = "1.3.3"
 borsh = "0.10.0"
+borsh1 = { package = "borsh", version = "1.5.7" }
 bv = "0.11.1"
 bytemuck = { version = "1.13.1", features = ["derive", "min_const_generics"] }
 cfg-if = "1.0.0"
 ed25519-dalek = "1.0.1"
 futures = "0.3.25"
-jito-priority-fee-distribution = { features = ["no-entrypoint"], git = "https://github.com/jito-foundation/jito-programs", branch = "tj/priority-fee-distributor" }
+jito-priority-fee-distribution = { features = ["no-entrypoint"], workspace = true }
 jito-steward = { features = ["no-entrypoint"], path = "../programs/steward" }
-jito-tip-distribution = { features = ["no-entrypoint"], git = "https://github.com/jito-foundation/jito-programs", branch = "tj/priority-fee-distributor" }
-jito-tip-distribution-sdk = { git = "https://github.com/jito-foundation/jito-programs", branch = "tj/priority-fee-distributor" }
+jito-tip-distribution = { features = ["no-entrypoint"], workspace = true }
+jito-tip-distribution-sdk = { workspace = true }
 rand = "0.7.3"
 rand_chacha = "0.3.1"
 semver = "1.0.17"
 serde = "1.0.183"
 serde_derive = "1.0.183"
-solana-gossip = "1.18"
-solana-program = "1.18"
-solana-program-test = "1.18"
-solana-sdk = "1.18"
-solana-version = "1.18"
-spl-associated-token-account = "2.1"
-spl-stake-pool = { features = ["no-entrypoint"], version = "1.0.0" }
+solana-gossip = "2.2"
+solana-program = "2.2"
+solana-program-test = "2.2"
+solana-sdk = "2.2"
+solana-version = "2.2"
+spl-associated-token-account = "7.0.0"
+spl-stake-pool = { features = ["no-entrypoint"], workspace = true }
 spl-token = "4.0"
 test-case = "3.3.1"
 validator-history = { features = ["no-entrypoint"], path = "../programs/validator-history" }
