+++ tests/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod priority_fee_distribution_helpers;
 mod spl_stake_pool_cli;
+pub mod stake_pool_utils;
 pub mod steward_fixtures;
 pub mod validator_history_fixtures;
