+++ utils/steward-cli/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 description = "CLI to manage the steward program"
 
 [dependencies]
-anchor-lang = "0.30.1"
+anchor-lang = "0.31.1"
 anyhow = "1.0.86"
 clap = { version = "4.3.0", features = ["derive", "env"] }
 dotenv = "0.15.0"
@@ -13,14 +13,14 @@ futures = "0.3.21"
 futures-util = "0.3.21"
 jito-steward = { features = ["no-entrypoint"], path = "../../programs/steward" }
 log = "0.4.18"
-solana-account-decoder = "1.18"
-solana-clap-utils = "1.18"
-solana-client = "1.18"
-solana-metrics = "1.18"
-solana-program = "1.18"
-solana-sdk = "1.18"
-spl-pod = "0.1.0"
-spl-stake-pool = { features = ["no-entrypoint"], version = "1.0.0" }
+solana-account-decoder = { workspace = true }
+solana-clap-utils = { workspace = true }
+solana-client = { workspace = true }
+solana-metrics = { workspace = true }
+solana-program = { workspace = true }
+solana-sdk = { workspace = true }
+spl-pod = { workspace = true }
+spl-stake-pool = { features = ["no-entrypoint"], workspace = true }
 stakenet-keeper = { path = "../../keepers/stakenet-keeper" }
 stakenet-sdk = { path = "../../sdk" }
 thiserror = "1.0.37"
