+++ utils/vote-state/Cargo.toml
@@ -7,7 +7,7 @@ license = "Apache-2.0"
 authors = ["Jito Foundation <team@jito.network>"]
 
 [dependencies]
-anchor-lang = "0.30.1"
+anchor-lang = "0.31.1"
 bincode = "1.3.3"
 serde = { version = "1.0.138", features = ["derive"] }
-solana-program = "1.18"
+solana-program = { workspace = true }
